등록금서비스 1월3일

+++ src/main/java/edu/du/campusflow/entity/TuitionTarget.java
@@ -28,7 +28,7 @@ public class TuitionTarget{
     private Member member;
 
     @Column(name = "paid_amount")
-    private Long paidAmount;
+    private Integer paidAmount;
 
     @Column(name = "paid_date")
     private LocalDateTime paidDate;
